+++ rl/policy/epsilon_greedy.py
@@ -1,23 +1,6 @@
 import numpy as np
+from rl.policy.base_policy import Policy
 from rl.util import log_self
-
-
-class Policy(object):
-
-    '''
-    The base class of Policy, with the core methods
-    '''
-
-    def __init__(self,
-                 **kwargs):  # absorb generic param without breaking
-        '''Construct externally, and set at Agent.compile()'''
-        self.agent = None
-
-    def select_action(self, state):
-        raise NotImplementedError()
-
-    def update(self, sys_vars):
-        raise NotImplementedError()
 
 
 class EpsilonGreedyPolicy(Policy):
@@ -57,7 +40,9 @@ class EpsilonGreedyPolicy(Policy):
         self.e = max(slope * epi + self.init_e, self.final_e)
         return self.e
 
+
 class DoubleDQNPolicy(EpsilonGreedyPolicy):
+
     '''
     Policy to accompany double dqn agents
     When actions are not random this policy
@@ -86,11 +71,14 @@ class DoubleDQNPolicy(EpsilonGreedyPolicy):
             action = np.argmax(Q_state)
         return action
 
+
 class DecayingEpsilonGreedyPolicy(EpsilonGreedyPolicy):
+
     '''
     Inspired by alvacarce's solution to mountain car
     https://gym.openai.com/evaluations/eval_t3GN2Xb0R5KpyjkJUGsLw
     '''
+
     def __init__(self,
                  init_e=1.0, final_e=0.1, exploration_anneal_episodes=30,
                  **kwargs):  # absorb generic param without breaking
@@ -102,6 +90,7 @@ class DecayingEpsilonGreedyPolicy(EpsilonGreedyPolicy):
         if self.e > self.final_e:
             self.e = self.e * self.decay
         return self.e
+
 
 class OscillatingEpsilonGreedyPolicy(EpsilonGreedyPolicy):
 
@@ -157,69 +146,3 @@ class TargetedEpsilonGreedyPolicy(EpsilonGreedyPolicy):
         if not active_exploration_cycle:
             self.e = max(self.e/2., self.final_e)
         return self.e
-
-
-class BoltzmannPolicy(Policy):
-
-    '''
-    The Boltzmann policy, where prob dist for selection
-    p = exp(Q/tau) / sum(Q[a]/tau)
-    '''
-
-    def __init__(self,
-                 init_tau=5., final_tau=0.5, exploration_anneal_episodes=20,
-                 **kwargs):  # absorb generic param without breaking
-        super(BoltzmannPolicy, self).__init__()
-        self.init_tau = init_tau
-        self.final_tau = final_tau
-        self.tau = self.init_tau
-        self.exploration_anneal_episodes = exploration_anneal_episodes
-        log_self(self)
-
-    def select_action(self, state):
-        agent = self.agent
-        state = np.reshape(state, (1, state.shape[0]))
-        Q_state = agent.model.predict(state)[0]  # extract from batch predict
-        assert Q_state.ndim == 1
-        Q_state = Q_state.astype('float64')  # fix precision nan issue
-        Q_state = Q_state - np.amax(Q_state)  # prevent overflow
-        exp_values = np.exp(Q_state / self.tau) + 0.001  # prevent underflow
-        assert not np.isnan(exp_values).any()
-        probs = np.array(exp_values / np.sum(exp_values))
-        probs /= probs.sum()  # renormalize to prevent floating pt error
-        action = np.random.choice(agent.env_spec['actions'], p=probs)
-        return action
-
-    def update(self, sys_vars):
-        '''strategy to update tau in agent'''
-        epi = sys_vars['epi']
-        rise = self.final_tau - self.init_tau
-        slope = rise / float(self.exploration_anneal_episodes)
-        self.tau = max(slope * epi + self.init_tau, self.final_tau)
-        return self.tau
-
-class DoubleDQNBoltzmannPolicy(BoltzmannPolicy):
-    '''
-    Same as the Boltzmann policy but for a Double DQN agent
-    '''
-
-    def __init__(self,
-                 init_tau=5., final_tau=0.5, exploration_anneal_episodes=20,
-                 **kwargs):  # absorb generic param without breaking
-        super(DoubleDQNBoltzmannPolicy, self).__init__()
-
-    def select_action(self, state):
-        agent = self.agent
-        state = np.reshape(state, (1, state.shape[0]))
-        Q_state1 = agent.model.predict(state)[0]  # extract from batch predict
-        Q_state2 = agent.model2.predict(state)[0]  # extract from batch predict
-        Q_state = Q_state1 + Q_state2
-        assert Q_state.ndim == 1
-        Q_state = Q_state.astype('float64')  # fix precision nan issue
-        Q_state = Q_state - np.amax(Q_state)  # prevent overflow
-        exp_values = np.exp(Q_state / self.tau) + 0.001  # prevent underflow
-        assert not np.isnan(exp_values).any()
-        probs = np.array(exp_values / np.sum(exp_values))
-        probs /= probs.sum()  # renormalize to prevent floating pt error
-        action = np.random.choice(agent.env_spec['actions'], p=probs)
-        return action
